// src/lib/polymarket/types.ts
export interface PolymarketMarket {
  id: string;
  question: string;
  conditionId: string;
  slug: string;
  resolutionSource?: string;
  endDate: string;
  liquidity?: number;
  startDate?: string;
  image?: string;
  icon?: string;
  description?: string;
  outcomes?: string;
  outcomePrices?: string;
  volume?: number;
  active?: boolean;
  closed?: boolean;
  archived?: boolean;
  new?: boolean;
  featured?: boolean;
  restricted?: boolean;
  groupItemTitle?: string;
  groupItemThreshold?: string;
  questionID?: string;
  enableOrderBook?: boolean;
  orderPriceMinTickSize?: number;
  orderMinSize?: number;
  volumeNum?: number;
  liquidityNum?: number;
  spread?: number;
  oneDayPriceChange?: number;
  lastTradePrice?: number;
  bestBid?: number;
  bestAsk?: number;
  automaticallyActive?: boolean;
  clobTokenIds?: string;
  events?: PolymarketEvent[];
}

export interface PolymarketEvent {
  id: string;
  ticker: string;
  slug: string;
  title: string;
  description?: string;
  resolutionSource?: string;
  startDate?: string;
  creationDate?: string;
  endDate: string;
  image?: string;
  icon?: string;
  active?: boolean;
  closed?: boolean;
  archived?: boolean;
  new?: boolean;
  featured?: boolean;
  restricted?: boolean;
  liquidity?: number;
  volume?: number;
  openInterest?: number;
  commentCount?: number;
  markets?: PolymarketMarket[];
  tags?: { id: string; label: string; slug: string }[];
  cyom?: boolean;
  enableOrderBook?: boolean;
  liquidityAmm?: number;
  liquidityClob?: number;
  volume24hr?: number;
  volume24hrAmm?: number;
  volume24hrClob?: number;
  volumeAmm?: number;
  volumeClob?: number;
  competitive?: number;
  series?: unknown[];
}